+++ ProjectData/GameSources/SpawnerBase.cpp
@@ -49,7 +49,7 @@ namespace basecross {
 	void SpawnerBase::ReadCSVLine(vector<wstring> csvLine, en_ObjectType objType) {
 		if (objType == en_ObjectType::en_Wave) {
 			for (size_t j = 0; j < csvLine.size(); j++) {
-				int waveTypeNum = stof(csvLine[j]);
+				int waveTypeNum = (int)stof(csvLine[j]);
 				float num = stof(csvLine[j + 1]);
 				m_waveTypeNum.push_back(waveTypeNum);
 				m_createPosX.push_back(num);
